+++ packages/good/lib/src/event/lifecycle.dart
@@ -166,7 +166,7 @@ mixin EntityLifecycleListener on GameListener {
 /// class SpatialIndexSystem extends GameSystem with EntitySpawnListener {
 ///   @override
 ///   void onEntitySpawned(Entity entity) {
-///     if (entity<Collider2D?>().component != null) index.insert(entity);
+///     if (entity.has<Collider2D>()) index.insert(entity);
 ///   }
 /// }
 /// ```